+++ testcontainers/Cargo.toml
@@ -32,24 +32,8 @@ watchdog = [ "signal-hook", "conquer-once" ]
 experimental = [ "async-trait", "bollard", "tokio" ]
 
 [dev-dependencies]
-aws-config = "0.49.0"
-aws-sdk-dynamodb = "0.19.0"
-aws-sdk-s3 = "0.19.0"
-aws-sdk-sqs = "0.19.0"
-aws-smithy-http = "0.49.0"
-aws-types = "0.49.0"
-bitcoincore-rpc = "0.16"
-json = "0.12"
-lapin = "1.8.0"
-mongodb = "2.1.0"
-orientdb-client = "0.6"
-postgres = "0.19"
 pretty_env_logger = "0.4"
-rdkafka = "0.29"
-redis = "0.22"
-reqwest = { version = "0.11", features = [ "blocking" ] }
-spectral = "0.6"
+reqwest = { version = "0.11.14", features = [ "blocking" ] }
+spectral = "0.6.0"
 testimages = { path = "../testimages" }
 tokio = { version = "1", features = [ "macros" ] }
-tokio-amqp = "1.0.0"
-zookeeper = "0.7"
