+++ testcontainers/src/images.rs
@@ -1,17 +1,1 @@
-pub mod coblox_bitcoincore;
-pub mod dynamodb_local;
-pub mod elastic_search;
-pub mod elasticmq;
 pub mod generic;
-pub mod google_cloud_sdk_emulators;
-pub mod hello_world;
-pub mod kafka;
-pub mod minio;
-pub mod mongo;
-pub mod orientdb;
-pub mod parity_parity;
-pub mod postgres;
-pub mod rabbitmq;
-pub mod redis;
-pub mod trufflesuite_ganachecli;
-pub mod zookeeper;
