--- testcontainers/src/images/hello_world.rs
@@ -1,20 +1,0 @@
-use crate::{core::WaitFor, Image};
-
-#[derive(Debug, Default)]
-pub struct HelloWorld;
-
-impl Image for HelloWorld {
-    type Args = ();
-
-    fn name(&self) -> String {
-        "hello-world".to_owned()
-    }
-
-    fn tag(&self) -> String {
-        "latest".to_owned()
-    }
-
-    fn ready_conditions(&self) -> Vec<WaitFor> {
-        vec![WaitFor::message_on_stdout("Hello from Docker!")]
-    }
-}
